requirements.txt
streamlit>=1.40.0
streamlit-folium>=0.22.0
folium>=0.17.0
leafmap>=0.38.0
requests>=2.31.0
httpx>=0.27.0
python-dotenv>=1.0.1
openai>=1.30.0
faiss-cpu>=1.8.0
numpy>=1.24.0
pyyaml>=6.0.1
